requirements.txt
# ============================================================================
# NFL Big Data Bowl 2026 - Python Dependencies
# ============================================================================
# 
# Install with: pip install -r requirements.txt
#
# ============================================================================

# Core Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Apache Airflow (if running locally without Docker)
apache-airflow==2.7.3

# Utilities
python-dateutil==2.8.2

# Optional: For development
ipython==8.18.1
jupyter==1.0.0

torch==2.0.1